Add test case to display how fake schema registry server fails to recognise identical schemas
Currently when invoking the POST `/subjects/:subject/versions` API in the FakeConfluentSchemaRegistryServer it fails to recognise when a schema being provided is identical to one already registrered in the fake registry. This is different to the actual confluent schema registry which if you send the same schema multiple times you receive the same id back each time.

This discrepency can cause problems for a user attempting to use the fake registry for testing purposes as eg they may rely on being able to assert that a kafka producer is invoked with the same encoded avro payload generated in their test case and in their application code. Because the fake registry currently does not recognise the two schemas used separately in the test case and in the application code invocation the assertion on the avro encoded payload will fail because the identifier in the message used to declare what the schema ID is will be different.

This commit also adds a regression test to ensure that when posting a repeated schema under different subjects it still returns a different identifier.